requirements.txt
certifi==2018.4.16
mysql-connector-python==2.2.3
numpy==1.13.1
protobuf==3.5.2
scipy==0.19.1
six==1.11.0
